fix(db): declare the attempts column bigint on Postgres
`test-pg` caught this: the column was declared `integer`, which is INT4 on
Postgres, while the row is decoded as `i64`:

    ColumnDecode { index: "attempts", source: "mismatched types; Rust type `i64`
    (as SQL type `INT8`) is not compatible with SQL type `INT4`" }

SQLite is untyped enough not to care, so the whole SQLite suite passed and the
gap only showed against a real Postgres.

The schema's own header states the rule — "INTEGER -> bigint (every integer column
is read as i64)" — so this just follows the convention the rest of the schema
already keeps.

+++ crates/cellarr-db/migrations/postgres/0013_import_attempt.sql
@@ -12,7 +12,7 @@
 -- `last_error` is kept so the reason is visible without trawling logs.
 CREATE TABLE IF NOT EXISTS import_attempt (
     grab_id      text PRIMARY KEY NOT NULL,
-    attempts     integer NOT NULL,
+    attempts     bigint NOT NULL,
     last_attempt text NOT NULL,
     last_error   text
 );
